HW13 is completed

+++ hw12_13_14_15_calendar/internal/storage/event.go
@@ -3,10 +3,10 @@ package storage
 import "time"
 
 type Event struct {
-	ID        int
+	ID        int64
 	Title     string
 	Descr     string
-	OwnerID   int
+	OwnerID   int64
 	StartDate time.Time
 	EndDate   time.Time
 	RemindFor time.Duration
